+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added support for oauth authentication using GitHub. ([#202](https://github.com/eclipse-csi/otterdog/issues/202))
 - Added support for auto-merging of PRs under certain conditions. ([#110](https://github.com/eclipse-csi/otterdog/issues/110))
 - Added handling for settings that require access to the Web UI. ([#208](https://github.com/eclipse-csi/otterdog/issues/208))
 - Added support for repository setting `private_vulnerability_reporting_enabled`. ([#205](https://github.com/eclipse-csi/otterdog/issues/205))